schoolsurvey/users/views.py
from rest_framework import viewsets
from rest_framework.permissions import AllowAny

from users.models import School
from users.serializers import SchoolSerializer
from users.permissions import IsLoggedInUserOrAdmin, IsAdminUser



class SchoolViewSet(viewsets.ModelViewSet):
    queryset = School.objects.all()
    serializer_class = SchoolSerializer
    
    def get_permissions(self):
        permission_classes = []
        if self.action == 'create':
            permission_classes = [AllowAny]
        elif self.action == 'retrieve' or self.action == 'update' or self.action == 'partial_update':
            permission_classes = [IsLoggedInUserOrAdmin]
        elif self.action == 'list' or self.action == 'destroy':
            permission_classes = [IsAdminUser]
        return [permission() for permission in permission_classes]