requirements.txt
Django==1.11
django-rest-auth==0.9.5
djangorestframework==3.8.2
Pillow==6.0.0
pytz==2019.1
six==1.12.0
